requirements.txt
asgiref==3.2.10
autopep8==1.5.4
braintree==4.3.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==2.9.2
Django==3.1
django-cors-headers==3.5.0
djangorestframework==3.11.1
idna==2.10
jwt==1.0.0
Markdown==3.2.2
Pillow==7.2.0
pycodestyle==2.6.0
pycparser==2.20
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
